[FIX] xlsx: remove useless normalization
The borders are already normalized at export time. The code was
denormalizing the renormalizing it without reason.

Task: 3733743
